+++ packages/hmssdk_flutter/ios/Classes/Models/HMSStreamingStateExtension.swift
@@ -15,11 +15,13 @@ class HMSStreamingStateExtension {
 
         dict["running"] = rtmp.running
         if let startedAt = rtmp.startedAt {
-            dict["started_at"] = "\(startedAt)"
+            dict["started_at"] = Int(startedAt.timeIntervalSince1970 * 1000)
         }
         if let error = rtmp.error {
             dict.merge(HMSErrorExtension.toDictionary(error)) { (_, new) in new }
         }
+        
+        dict["state"] = rtmp.state.displayString().uppercased()
         return dict
     }
 
@@ -29,12 +31,13 @@ class HMSStreamingStateExtension {
 
         dict["running"] = server.running
         if let startedAt = server.startedAt {
-            dict["started_at"] = "\(startedAt)"
+            dict["started_at"] = Int(startedAt.timeIntervalSince1970 * 1000)
         }
         if let error = server.error {
             dict.merge(HMSErrorExtension.toDictionary(error)) { (_, new) in new }
         }
-
+        
+        dict["state"] = server.state.displayString().uppercased()
         return dict
     }
 
@@ -46,12 +49,13 @@ class HMSStreamingStateExtension {
 
         dict["running"] = browser.running
         if let startedAt = browser.startedAt {
-            dict["started_at"] = "\(startedAt)"
+            dict["started_at"] = Int(startedAt.timeIntervalSince1970 * 1000)
         }
         if let error = browser.error {
             dict.merge(HMSErrorExtension.toDictionary(error)) { (_, new) in new }
         }
 
+        dict["state"] = browser.state.displayString().uppercased()
         return dict
     }
 
@@ -65,6 +69,7 @@ class HMSStreamingStateExtension {
         }
         dict["variants"]=args
 
+        dict["state"] = hlsStreaming.state.displayString().uppercased()
         return dict
     }
 
@@ -73,12 +78,13 @@ class HMSStreamingStateExtension {
 
         dict["running"] = hlsRecording.running
         if let startedAt = hlsRecording.startedAt {
-            dict["started_at"] = "\(startedAt)"
+            dict["started_at"] = Int(startedAt.timeIntervalSince1970 * 1000)
         }
         if let error = hlsRecording.error {
             dict.merge(HMSErrorExtension.toDictionary(error)) { (_, new) in new }
         }
 
+        dict["state"] = hlsRecording.state.displayString().uppercased()
         return dict
     }
 }